sdxl-turbo/requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
awscli==1.32.14
boto3==1.34.14
fastapi==0.105.0
Pillow==10.2.0
pybase64==1.3.2
pydantic==2.5.3
uvicorn==0.20.0
diffusers==0.25.0
transformers==4.36.2 
